Refactor router to not mix routing strategies in one class
