+++ packages/ddc/src/Router/BaseStrategy.web.ts
@@ -1,9 +1,9 @@
-import { BaseRouter, RouterOperation, RouterNode } from './BaseRouter';
+import { RouterNode, RouterOperation, RoutingStrategy } from './RoutingStrategy';
 
 const isSSLRequired = () => globalThis.location?.protocol === 'https:';
 
-export class Router extends BaseRouter {
-  selectNode(operation: RouterOperation, nodes: RouterNode[]): RouterNode {
+export abstract class BaseStrategy extends RoutingStrategy {
+  selectNode(operation: RouterOperation, nodes: RouterNode[]) {
     const sslNodes = isSSLRequired() ? nodes.filter((node) => node.ssl) : nodes;
     const randomIndex = Math.floor(Math.random() * sslNodes.length);
